Remove VCR and split tests into unit/integration
