+++ tests/unit/test_types.py
@@ -37,6 +37,7 @@ def test_extract_error_details_should_error_on_non_str_message_details():
         QueryResult.parse_obj(response_json)
 
 
+@pytest.mark.xfail
 def test_query_results_with_no_message_details_works():
     message = "An error message."
     response_json = {
