--- tests/vcr/cassettes/fixtures/looker_client.yaml
@@ -1,139 +1,0 @@
-interactions:
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '69'
-      content-type:
-      - application/x-www-form-urlencoded
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: POST
-    uri: https://spectacles.looker.com:19999/api/3.1/login
-  response:
-    content: '{"access_token": "<Access token filtered from cassette>", "token_type":
-      "Bearer", "expires_in": 3600}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 00:10:08 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 4be96def395c9ba6
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 4be96def395c9ba6
-      X-B3-TraceId:
-      - 62aa74e08cd9c7144be96def395c9ba6
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/versions
-  response:
-    content: '{"looker_release_version":"22.8.32","current_version":{"version":"4.0","full_version":"4.0.22.8","status":"current","swagger_url":"https://spectacles.looker.com/api/4.0/swagger.json"},"supported_versions":[{"version":"2.99","full_version":"2.99.0","status":"internal_test","swagger_url":"https://spectacles.looker.com/api/2.99/swagger.json"},{"version":"3.0","full_version":"3.0.0","status":"legacy","swagger_url":"https://spectacles.looker.com/api/3.0/swagger.json"},{"version":"3.1","full_version":"3.1.0","status":"legacy","swagger_url":"https://spectacles.looker.com/api/3.1/swagger.json"},{"version":"4.0","full_version":"4.0.22.8","status":"current","swagger_url":"https://spectacles.looker.com/api/4.0/swagger.json"}],"api_server_url":"https://spectacles.looker.com","web_server_url":"https://spectacles.looker.com"}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 00:10:08 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 8e3ec23217683fd5
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 8e3ec23217683fd5
-      X-B3-TraceId:
-      - 62aa74e033f684ee8e3ec23217683fd5
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: '{"workspace_id": "production"}'
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '30'
-      content-type:
-      - application/json
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: PATCH
-    uri: https://spectacles.looker.com:19999/api/3.1/session
-  response:
-    content: '{"workspace_id":"production","sudo_user_id":null,"can":{"view":true,"update":true}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 00:10:08 GMT
-      Set-Cookie:
-      - looker.browser=59290397; expires=Sun, 15 Jun 2025 00:10:08 GMT; HttpOnly
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - e9b1dff0fdff6b56
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - e9b1dff0fdff6b56
-      X-B3-TraceId:
-      - 62aa74e0b5cf5c18e9b1dff0fdff6b56
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-version: 1
