--- tests/vcr/cassettes/test_branch_manager_vcr/test_manage_prod_with_advanced_deploy.yaml
@@ -1,226 +1,0 @@
-interactions:
-- request:
-    body: '{"workspace_id": "production"}'
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '30'
-      content-type:
-      - application/json
-      cookie:
-      - looker.browser=78847301
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: PATCH
-    uri: https://spectacles.looker.com:19999/api/3.1/session
-  response:
-    content: '{"workspace_id":"production","sudo_user_id":null,"can":{"view":true,"update":true}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 20:25:28 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 35f060425aadc5c6
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 35f060425aadc5c6
-      X-B3-TraceId:
-      - 62ab91b81d37fa6735f060425aadc5c6
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      cookie:
-      - looker.browser=78847301
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/projects/spectacles-advanced-deploy/git_branch
-  response:
-    content: '{"name":"master","remote":"origin","remote_name":"master","error":null,"message":null,"owner_name":null,"readonly":false,"personal":false,"is_local":true,"is_remote":true,"is_production":true,"ahead_count":0,"behind_count":0,"commit_at":1639134522,"ref":"0af495920bf685b83fd72207056c4fb7e09cf347","remote_ref":"0af495920bf685b83fd72207056c4fb7e09cf347","can":{}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 20:25:28 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 0217af208b256c72
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 0217af208b256c72
-      X-B3-TraceId:
-      - 62ab91b84b7871210217af208b256c72
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      cookie:
-      - looker.browser=78847301
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/session
-  response:
-    content: '{"workspace_id":"production","sudo_user_id":null,"can":{"view":true,"update":true}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 20:25:28 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 95b7d979aadfc42a
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 95b7d979aadfc42a
-      X-B3-TraceId:
-      - 62ab91b8c074105795b7d979aadfc42a
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      cookie:
-      - looker.browser=78847301
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/projects/spectacles-advanced-deploy/git_branch
-  response:
-    content: '{"name":"master","remote":"origin","remote_name":"master","error":null,"message":null,"owner_name":null,"readonly":false,"personal":false,"is_local":true,"is_remote":true,"is_production":true,"ahead_count":0,"behind_count":0,"commit_at":1639134522,"ref":"0af495920bf685b83fd72207056c4fb7e09cf347","remote_ref":"0af495920bf685b83fd72207056c4fb7e09cf347","can":{}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 20:25:28 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 8c7e8f877e5ff075
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 8c7e8f877e5ff075
-      X-B3-TraceId:
-      - 62ab91b83a51b1988c7e8f877e5ff075
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      cookie:
-      - looker.browser=78847301
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/session
-  response:
-    content: '{"workspace_id":"production","sudo_user_id":null,"can":{"view":true,"update":true}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 20:25:28 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 872e9441c1ad2536
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 872e9441c1ad2536
-      X-B3-TraceId:
-      - 62ab91b8632d324e872e9441c1ad2536
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-version: 1
