--- tests/vcr/cassettes/test_branch_manager_vcr/test_manage_with_ref_not_present_in_local_repo.yaml
@@ -1,164 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - PyGithub/Python
-    method: GET
-    uri: https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md?ref=master
-  response:
-    body:
-      string: '{"name":"README.md","path":"README.md","sha":"025fc068587560d0ab40ba3b90508d8ea0f1e1a8","size":37,"url":"https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md?ref=master","html_url":"https://github.com/spectacles-ci/eye-exam/blob/master/README.md","git_url":"https://api.github.com/repos/spectacles-ci/eye-exam/git/blobs/025fc068587560d0ab40ba3b90508d8ea0f1e1a8","download_url":"https://raw.githubusercontent.com/spectacles-ci/eye-exam/master/README.md?token=ACCFHKYLHWX6ZYOLN3JPW4LCVOI7K","type":"file","content":"IyBFeWUgRXhhbQoKVGhpcyBuZWVkcyB0byBiZSBjaGFuZ2VkCg==\n","encoding":"base64","_links":{"self":"https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md?ref=master","git":"https://api.github.com/repos/spectacles-ci/eye-exam/git/blobs/025fc068587560d0ab40ba3b90508d8ea0f1e1a8","html":"https://github.com/spectacles-ci/eye-exam/blob/master/README.md"}}'
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Access-Control-Expose-Headers:
-      - ETag, Link, Location, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining,
-        X-RateLimit-Used, X-RateLimit-Resource, X-RateLimit-Reset, X-OAuth-Scopes,
-        X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type, X-GitHub-SSO,
-        X-GitHub-Request-Id, Deprecation, Sunset
-      Cache-Control:
-      - private, max-age=60, s-maxage=60
-      Content-Security-Policy:
-      - default-src 'none'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Thu, 16 Jun 2022 20:25:29 GMT
-      ETag:
-      - W/"025fc068587560d0ab40ba3b90508d8ea0f1e1a8"
-      Last-Modified:
-      - Mon, 06 Jun 2022 10:11:15 GMT
-      Referrer-Policy:
-      - origin-when-cross-origin, strict-origin-when-cross-origin
-      Server:
-      - GitHub.com
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubdomains; preload
-      Transfer-Encoding:
-      - chunked
-      Vary:
-      - Accept, Authorization, Cookie, X-GitHub-OTP
-      - Accept-Encoding, Accept, X-Requested-With
-      X-Accepted-OAuth-Scopes:
-      - ''
-      X-Content-Type-Options:
-      - nosniff
-      X-Frame-Options:
-      - deny
-      X-GitHub-Media-Type:
-      - github.v3; format=json
-      X-GitHub-Request-Id:
-      - CE0A:135B:6EFA3A:D4CB2D:62AB91B8
-      X-OAuth-Scopes:
-      - repo
-      X-RateLimit-Limit:
-      - '5000'
-      X-RateLimit-Remaining:
-      - '4925'
-      X-RateLimit-Reset:
-      - '1655413239'
-      X-RateLimit-Resource:
-      - core
-      X-RateLimit-Used:
-      - '75'
-      X-XSS-Protection:
-      - '0'
-      content-length:
-      - '894'
-      github-authentication-token-expiration:
-      - 2022-09-14 16:54:58 UTC
-    status:
-      code: 200
-      message: OK
-- request:
-    body: '{"message": "Updating file", "content": "Lg==", "sha": "025fc068587560d0ab40ba3b90508d8ea0f1e1a8",
-      "branch": "pytest-tmp"}'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '122'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - PyGithub/Python
-    method: PUT
-    uri: https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md
-  response:
-    body:
-      string: '{"content":{"name":"README.md","path":"README.md","sha":"945c9b46d684f08ec84cb316e1dc0061e361f794","size":1,"url":"https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md?ref=pytest-tmp","html_url":"https://github.com/spectacles-ci/eye-exam/blob/pytest-tmp/README.md","git_url":"https://api.github.com/repos/spectacles-ci/eye-exam/git/blobs/945c9b46d684f08ec84cb316e1dc0061e361f794","download_url":"https://raw.githubusercontent.com/spectacles-ci/eye-exam/pytest-tmp/README.md?token=ACCFHKZVNXXAFMAE3EJZ6I3CVOI7K","type":"file","_links":{"self":"https://api.github.com/repos/spectacles-ci/eye-exam/contents/README.md?ref=pytest-tmp","git":"https://api.github.com/repos/spectacles-ci/eye-exam/git/blobs/945c9b46d684f08ec84cb316e1dc0061e361f794","html":"https://github.com/spectacles-ci/eye-exam/blob/pytest-tmp/README.md"}},"commit":{"sha":"f6733093ab97104778a0162ce90cf44bca384e72","node_id":"C_kwDODcbQgdoAKGY2NzMzMDkzYWI5NzEwNDc3OGEwMTYyY2U5MGNmNDRiY2EzODRlNzI","url":"https://api.github.com/repos/spectacles-ci/eye-exam/git/commits/f6733093ab97104778a0162ce90cf44bca384e72","html_url":"https://github.com/spectacles-ci/eye-exam/commit/f6733093ab97104778a0162ce90cf44bca384e72","author":{"name":"Josh
-        Temple","email":"8672171+joshtemple@users.noreply.github.com","date":"2022-06-16T20:25:29Z"},"committer":{"name":"Josh
-        Temple","email":"8672171+joshtemple@users.noreply.github.com","date":"2022-06-16T20:25:29Z"},"tree":{"sha":"e576f27cc5d17de6340b93eb3edcfeb6bd41b9d6","url":"https://api.github.com/repos/spectacles-ci/eye-exam/git/trees/e576f27cc5d17de6340b93eb3edcfeb6bd41b9d6"},"message":"Updating
-        file","parents":[{"sha":"785a58bed86f4891f3b63de17f33ec8fab0ffdad","url":"https://api.github.com/repos/spectacles-ci/eye-exam/git/commits/785a58bed86f4891f3b63de17f33ec8fab0ffdad","html_url":"https://github.com/spectacles-ci/eye-exam/commit/785a58bed86f4891f3b63de17f33ec8fab0ffdad"}],"verification":{"verified":false,"reason":"unsigned","signature":null,"payload":null}}}'
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Access-Control-Expose-Headers:
-      - ETag, Link, Location, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining,
-        X-RateLimit-Used, X-RateLimit-Resource, X-RateLimit-Reset, X-OAuth-Scopes,
-        X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type, X-GitHub-SSO,
-        X-GitHub-Request-Id, Deprecation, Sunset
-      Cache-Control:
-      - private, max-age=60, s-maxage=60
-      Content-Security-Policy:
-      - default-src 'none'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Thu, 16 Jun 2022 20:25:29 GMT
-      ETag:
-      - W/"7d6024ed6ea318227a388587fd913b468588ca5b6e7e43b5ac7322e5cb7daea3"
-      Referrer-Policy:
-      - origin-when-cross-origin, strict-origin-when-cross-origin
-      Server:
-      - GitHub.com
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubdomains; preload
-      Transfer-Encoding:
-      - chunked
-      Vary:
-      - Accept, Authorization, Cookie, X-GitHub-OTP
-      - Accept-Encoding, Accept, X-Requested-With
-      X-Accepted-OAuth-Scopes:
-      - ''
-      X-Content-Type-Options:
-      - nosniff
-      X-Frame-Options:
-      - deny
-      X-GitHub-Media-Type:
-      - github.v3; format=json
-      X-GitHub-Request-Id:
-      - CE0A:135B:6EFA48:D4CB43:62AB91B9
-      X-OAuth-Scopes:
-      - repo
-      X-RateLimit-Limit:
-      - '5000'
-      X-RateLimit-Remaining:
-      - '4924'
-      X-RateLimit-Reset:
-      - '1655413239'
-      X-RateLimit-Resource:
-      - core
-      X-RateLimit-Used:
-      - '76'
-      X-XSS-Protection:
-      - '0'
-      content-length:
-      - '1991'
-      github-authentication-token-expiration:
-      - 2022-09-14 16:54:58 UTC
-    status:
-      code: 200
-      message: OK
-version: 1
