--- tests/vcr/cassettes/test_client_vcr/test_bad_authentication_request_should_raise_looker_api_error.yaml
@@ -1,68 +1,0 @@
-interactions:
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '59'
-      content-type:
-      - application/x-www-form-urlencoded
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: POST
-    uri: https://spectacles.looker.com:19999/api/3.1/login
-  response:
-    content: "<!DOCTYPE html>\n<html>\n<head>\n    <title>Looker Not Found (404)</title>\n
-      \   <link href='https://fonts.googleapis.com/css?family=Open+Sans:400,600' rel='stylesheet'
-      type='text/css'>\n\n    <!-- @@@@@@@@@@@@@ FAVICONS @@@@@@@@@@@@@ -->\n\n    <link
-      rel=\"apple-touch-icon-precomposed\" sizes=\"57x57\" href=\"https://wwwstatic-b.lookercdn.com/favicon/apple-touch-icon-57x57.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"114x114\" href=\"https://wwwstatic-c.lookercdn.com/favicon/apple-touch-icon-114x114.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"72x72\" href=\"https://wwwstatic-d.lookercdn.com/favicon/apple-touch-icon-72x72.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"144x144\" href=\"https://wwwstatic-a.lookercdn.com/favicon/apple-touch-icon-144x144.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"60x60\" href=\"https://wwwstatic-b.lookercdn.com/favicon/apple-touch-icon-60x60.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"120x120\" href=\"https://wwwstatic-c.lookercdn.com/favicon/apple-touch-icon-120x120.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"76x76\" href=\"https://wwwstatic-d.lookercdn.com/favicon/apple-touch-icon-76x76.png\"
-      />\n    <link rel=\"apple-touch-icon-precomposed\" sizes=\"152x152\" href=\"https://wwwstatic-a.lookercdn.com/favicon/apple-touch-icon-152x152.png\"
-      />\n    <link rel=\"icon\" type=\"image/png\" href=\"https://wwwstatic-b.lookercdn.com/favicon/favicon-196x196.png\"
-      sizes=\"196x196\" />\n    <link rel=\"icon\" type=\"image/png\" href=\"https://wwwstatic-c.lookercdn.com/favicon/favicon-96x96.png\"
-      sizes=\"96x96\" />\n    <link rel=\"icon\" type=\"image/png\" href=\"https://wwwstatic-d.lookercdn.com/favicon/favicon-32x32.png\"
-      sizes=\"32x32\" />\n    <link rel=\"icon\" type=\"image/png\" href=\"https://wwwstatic-a.lookercdn.com/favicon/favicon-16x16.png\"
-      sizes=\"16x16\" />\n    <link rel=\"icon\" type=\"image/png\" href=\"https://wwwstatic-b.lookercdn.com/favicon/favicon-128.png\"
-      sizes=\"128x128\" />\n    <meta name=\"application-name\" content=\"Looker\"/>\n
-      \   <meta name=\"msapplication-TileColor\" content=\"#FFFFFF\" />\n    <meta
-      name=\"msapplication-TileImage\" content=\"https://wwwstatic-c.lookercdn.com/favicon/mstile-144x144.png\"
-      />\n    <meta name=\"msapplication-square70x70logo\" content=\"https://wwwstatic-d.lookercdn.com/favicon/mstile-70x70.png\"
-      />\n    <meta name=\"msapplication-square150x150logo\" content=\"https://wwwstatic-a.lookercdn.com/favicon/mstile-150x150.png\"
-      />\n    <meta name=\"msapplication-wide310x150logo\" content=\"https://wwwstatic-b.lookercdn.com/favicon/mstile-310x150.png\"
-      />\n    <meta name=\"msapplication-square310x310logo\" content=\"https://wwwstatic-c.lookercdn.com/favicon/mstile-310x310.png\"
-      />\n\n    <style type=\"text/css\">\n        body {\n            background-color:
-      #2e343f;\n            color: white;\n            height: auto;\n            font-family:
-      Open Sans, Helvetica, Arial, sans-serif;\n        }\n        .message {\n            width:
-      100%;\n            max-width: 760px;\n            margin: 0 auto;\n            margin-top:
-      135px;\n            text-align: center;\n        }\n        h2, h3 {\n            font-weight:
-      normal;\n        }\n        a {\n            color: white;\n        }\n    </style>\n</head>\n<body>\n\n
-      \   <div class=\"message\">\n\n        <img width=\"210\" height=\"84\" src=\"https://wwwstatic-a.lookercdn.com/logos/looker_all_white.svg\"
-      alt=\"Looker\">\n\n        <h1>Looker is unavailable.</h1>\n\n        <h2>If
-      you typed in a URL, double-check the spelling.</h2>\n        <h2>This may also
-      be due to a temporary condition such as an outage, <a href=\"https://docs.looker.com/relnotes/hosted-maintenance-hours\">scheduled
-      maintenance</a> or upgrade.</h2>\n        <br>\n        <h3>\n          If this
-      message persists or you have any concerns, <br> contact us from\n          <a
-      href=\"https://help.looker.com\">help.looker.com</a> and we'll respond promptly.\n
-      \       </h3>\n\n    </div>\n\n</body>\n</html>\n"
-    headers:
-      Connection:
-      - close
-      Content-Type:
-      - '*/*'
-      Date:
-      - Thu, 16 Jun 2022 00:10:21 GMT
-    http_version: HTTP/1.1
-    status_code: 404
-version: 1
