--- tests/vcr/cassettes/test_client_vcr/test_create_query_without_dimensions_should_return_certain_fields.yaml
@@ -1,51 +1,0 @@
-interactions:
-- request:
-    body: '{"model": "eye_exam", "view": "users", "fields": [], "limit": 0, "filter_expression":
-      "1=2"}'
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '92'
-      content-type:
-      - application/json
-      cookie:
-      - looker.browser=59290397
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: POST
-    uri: https://spectacles.looker.com:19999/api/3.1/queries?fields=
-  response:
-    content: '{"id":1945,"view":"users","fields":null,"pivots":null,"fill_fields":null,"filters":null,"filter_expression":"1=2","sorts":null,"limit":"0","column_limit":null,"total":null,"row_total":null,"subtotals":null,"vis_config":null,"filter_config":null,"visible_ui_sections":null,"slug":"64n322h","client_id":"DIXL8LS20QeCUNhDRkywY3","share_url":"https://spectacles.looker.com/x/DIXL8LS20QeCUNhDRkywY3","expanded_share_url":"https://spectacles.looker.com/explore/eye_exam/users?filter_expression=1%3D2\u0026limit=0\u0026origin=share-expanded","url":"/explore/eye_exam/users?filter_expression=1%3D2\u0026limit=0","runtime":null,"has_table_calculations":false,"model":"eye_exam","dynamic_fields":null,"query_timezone":null,"quick_calcs":null,"analysis_config":null,"can":{"run":true,"see_results":true,"explore":true,"create":true,"show":true,"cost_estimate":true,"index":true,"see_lookml":true,"see_aggregate_table_lookml":true,"see_derived_table_lookml":true,"see_sql":true,"save":true,"generate_drill_links":true,"download":true,"download_unlimited":true,"use_custom_fields":true,"edit_custom_fields":true,"edit_table_calculations":true,"schedule":true,"render":true}}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 00:10:21 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 3bd7027909c0c5d9
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 3bd7027909c0c5d9
-      X-B3-TraceId:
-      - 62aa74ed72763b413bd7027909c0c5d9
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-version: 1
