--- tests/vcr/cassettes/test_client_vcr/test_get_looker_release_version_should_return_correct_version.yaml
@@ -1,44 +1,0 @@
-interactions:
-- request:
-    body: ''
-    headers:
-      accept:
-      - '*/*'
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      host:
-      - spectacles.looker.com:19999
-      user-agent:
-      - python-httpx/0.23.0
-    method: GET
-    uri: https://spectacles.looker.com:19999/api/3.1/versions
-  response:
-    content: '{"looker_release_version":"22.8.32","current_version":{"version":"4.0","full_version":"4.0.22.8","status":"current","swagger_url":"https://spectacles.looker.com/api/4.0/swagger.json"},"supported_versions":[{"version":"2.99","full_version":"2.99.0","status":"internal_test","swagger_url":"https://spectacles.looker.com/api/2.99/swagger.json"},{"version":"3.0","full_version":"3.0.0","status":"legacy","swagger_url":"https://spectacles.looker.com/api/3.0/swagger.json"},{"version":"3.1","full_version":"3.1.0","status":"legacy","swagger_url":"https://spectacles.looker.com/api/3.1/swagger.json"},{"version":"4.0","full_version":"4.0.22.8","status":"current","swagger_url":"https://spectacles.looker.com/api/4.0/swagger.json"}],"api_server_url":"https://spectacles.looker.com","web_server_url":"https://spectacles.looker.com"}'
-    headers:
-      Connection:
-      - close
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Thu, 16 Jun 2022 00:10:21 GMT
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-      Vary:
-      - Accept-Encoding, Origin
-      X-B3-ParentSpanId:
-      - 3d1b7415b3c5ba84
-      X-B3-Sampled:
-      - '0'
-      X-B3-SpanId:
-      - 3d1b7415b3c5ba84
-      X-B3-TraceId:
-      - 62aa74ed1594bca03d1b7415b3c5ba84
-      X-Content-Type-Options:
-      - nosniff
-    http_version: HTTP/1.1
-    status_code: 200
-version: 1
